tests: Fix -Wformat gcc warning
test_fork_thread_deadlock.c:48:25: warning: format ‘%d’ expects argument
    of type ‘int’, but argument 3 has type ‘pid_t {aka long int}’ [-Wformat=]

+++ tests/test_fork_thread_deadlock.c
@@ -45,7 +45,7 @@
 
 static void test_swrap_signal_handler(int signum)
 {
-	fprintf(stderr, "PID: %d, SIGNUM: %d\n", getpid(), signum);
+	fprintf(stderr, "PID: %u, SIGNUM: %d\n", (unsigned int)getpid(), signum);
 	write(1, "DEADLOCK?\n", 10);
 }
